+++ Example/Pods/Target Support Files/Highlightr-OSX/Highlightr-OSX-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Highlightr_OSX : NSObject
-@end
-@implementation PodsDummy_Highlightr_OSX
-@end