+++ Example/Pods/Target Support Files/Pods-Highlightr_OSX_Example/Pods-Highlightr_OSX_Example-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_Highlightr_OSX_Example : NSObject
-@end
-@implementation PodsDummy_Pods_Highlightr_OSX_Example
-@end